add proptypes and separate container and component
